+++ sei-ibc-go/testing/chain.go
@@ -72,9 +72,9 @@ type TestChain struct {
 	SenderAccounts []SenderAccount
 }
 
-// NewTestChain initializes a new TestChain instance with a single validator set using a
-// generated secp256k1 Tendermint private key. It also creates a sender BaseAccount to be used for
-// delivering transactions.
+// NewTestChainWithValSet initializes a new TestChain instance with the given validator set
+// and signer array. It also initializes 10 Sender accounts with a balance of 10000000000000000000 coins of
+// bond denom to use for tests.
 //
 // The first block height is committed to state in order to allow for client creations on
 // counterparty chains. The TestChain will return with a block height starting at 2.
@@ -84,17 +84,11 @@ type TestChain struct {
 //
 // NOTE: to use a custom sender privkey and account for testing purposes, replace and modify this
 // constructor function.
-func NewTestChain(t *testing.T, coord *Coordinator, chainID string) *TestChain {
-	// generate validator private/public key
-	privVal := mock.NewPV()
-	pubKey, err := privVal.GetPubKey()
-	require.NoError(t, err)
-
-	// create validator set with single validator
-	validator := tmtypes.NewValidator(pubKey, 1)
-	valSet := tmtypes.NewValidatorSet([]*tmtypes.Validator{validator})
-	signers := []tmtypes.PrivValidator{privVal}
-
+//
+// CONTRACT: Validator and signer array must be provided in the order expected by Tendermint.
+// i.e. sorted first by power and then lexicographically by address.
+func NewTestChainWithValSet(t *testing.T, coord *Coordinator, chainID string, valSet *tmtypes.ValidatorSet, signers []tmtypes.PrivValidator) *TestChain {
+  
 	genAccs := []authtypes.GenesisAccount{}
 	genBals := []banktypes.Balance{}
 	senderAccs := []SenderAccount{}
@@ -153,6 +147,38 @@ func NewTestChain(t *testing.T, coord *Coordinator, chainID string) *TestChain {
 	coord.CommitBlock(chain)
 
 	return chain
+}
+
+// NewTestChain initializes a new test chain with a default of 4 validators
+// Use this function if the tests do not need custom control over the validator set
+func NewTestChain(t *testing.T, coord *Coordinator, chainID string) *TestChain {
+	// generate validators private/public key
+	var (
+		validatorsPerChain = 4
+		validators         []*tmtypes.Validator
+		signersByAddress   = make(map[string]tmtypes.PrivValidator, validatorsPerChain)
+	)
+
+	for i := 0; i < validatorsPerChain; i++ {
+		privVal := mock.NewPV()
+		pubKey, err := privVal.GetPubKey()
+		require.NoError(t, err)
+		validators = append(validators, tmtypes.NewValidator(pubKey, 1))
+		signersByAddress[pubKey.Address().String()] = privVal
+	}
+
+	// construct validator set;
+	// Note that the validators are sorted by voting power
+	// or, if equal, by address lexical order
+	valSet := tmtypes.NewValidatorSet(validators)
+
+	// create signers indexed by the valSet validators's order
+	signers := []tmtypes.PrivValidator{}
+	for _, val := range valSet.Validators {
+		signers = append(signers, signersByAddress[val.PubKey.Address().String()])
+	}
+
+	return NewTestChainWithValSet(t, coord, chainID, valSet, signers)
 }
 
 // GetContext returns the current context for the application.